consolidate previous commit into pre-existing moderation.tsx
