+++ src/lib/misc.ts
@@ -22,31 +22,6 @@ import * as preferences from "preferences";
 import { alert } from "swal_config";
 import React from "react";
 
-// Must match back-end MODERATOR_POWER
-
-export enum MODERATOR_POWERS {
-    NONE = 0,
-    HANDLE_SCORE_CHEAT = 0b001,
-    HANDLE_ESCAPING = 0b010,
-    HANDLE_STALLING = 0b100,
-}
-
-export const MOD_POWER_NAMES: { [key in MODERATOR_POWERS]: string } = {
-    [MODERATOR_POWERS.NONE]: pgettext("... as in 'moderators powers: None'", "None"),
-    [MODERATOR_POWERS.HANDLE_SCORE_CHEAT]: pgettext(
-        "A label for a moderator power",
-        "Handle Score Cheating Reports",
-    ),
-    [MODERATOR_POWERS.HANDLE_ESCAPING]: pgettext(
-        "A label for a moderator power",
-        "Handle Escaping Reports",
-    ),
-    [MODERATOR_POWERS.HANDLE_STALLING]: pgettext(
-        "A label for a moderator power",
-        "Handle Stalling Reports",
-    ),
-};
-
 export type Timeout = ReturnType<typeof setTimeout>;
 
 export function updateDup(obj: any, field: string, value: any) {
